simplify(#178): extract lang_map from format.rs to meet 400-line limit
- Extract lang_to_id/lang_from_id into index/lang_map.rs with dedicated
  lang_map_tests.rs, bringing format.rs from 514 → 400 lines
- Introduce read_array<N> helper to eliminate 7 repetitive try_into+map_err
  patterns in decode_header/decode_entry/decode_posting/decode_file_meta
- Merge "Format constants" and "BM25 parameters" section banners into one
  "Constants" section; inline BM25_K1/BM25_B as private constants
- Merge "Checksum" and "IDF helper" sections into "BM25 scoring and checksum"
- All 113 tests pass, zero warnings
