+++ crates/rskim-search/src/index/mod.rs
@@ -29,6 +29,7 @@
 
 mod builder;
 mod format;
+mod lang_map;
 mod reader;
 
 pub use builder::NgramIndexBuilder;
